Rename get_changed_fuzzers_since_last and move to own module.
New name should be clearer and less verbose.
Move to it's own module so that change_utils doesn't need db code.
